fix(docker): add dedicated volume for node_modules
Creating a shared volume with ".:/app" also shares the node_modules between the host and the container.
This is problematic because vite creates a cache folder ".vite" in the node_modules with docker's root permissions.
Thus, vite on the host machine fails to launch as it tries to access this cache.
The new volume isolates the node_modules of the container from the host's, avoiding the problem altogether.